workaround to check/fix users stuck on pistar-update v3.3 or v3.4
This is a purposed workaround to check/fix users stuck on pistar-update v3.3 or v3.4.
NOTE: if you agree and merge this then you MUST 1st change current version of the update script to v3.5 !!

Why here on platformDetect.sh? This script is part of the Bin repo, and all users should be able to update that repo, even the ones with the loop problem, as this is the 1st repo to be updated - before entering the loop. Also this script runs when user just visits admin or configure pages, then... if user tries to Update and get stuck on the loop then if he go back to admin or configure page this workaround will run and next time he tries to Update it will be all sorted.

What do you think?